code refactoring

+++ Assets/Scripts/IEnemy.cs
@@ -3,6 +3,6 @@
     string Name { get; set; }
     float Hitpoints { get; set; }
     int Damage { get; set; }
-    int FundAmount { get; set; }
+    float FundAmount { get; set; }
     void Die();
 }
